added streaming parsing

+++ lib/rafflesstore.php
@@ -1,13 +1,17 @@
 <?php
 require_once 'ldpath.php';
 require_once 'index.php';
+require_once 'hierarchicalindex.php';
 require_once 'descriptionstore.php';
+require_once 'streamingparsers.php';
 
 class RafflesStore {
   var $Index;
+  var $HierarchicalIndex;
   var $DescriptionStore;
   var $dirname;
   var $LDPath;
+  var $_lastSet=array();
   var $prefixes = array(
     "foaf" =>	"http://xmlns.com/foaf/0.1/",
     "rdf" =>	"http://www.w3.org/1999/02/22-rdf-syntax-ns#",
@@ -41,9 +45,10 @@ class RafflesStore {
     }
     $this->DescriptionStore = new DescriptionStore($dirname . DIRECTORY_SEPARATOR . 'descriptions');
     $this->LDPath = new LDPath($this->prefixes);
+    $this->HierarchicalIndex = new HierarchicalIndex($this->Index);
   }
 
-  function addNamespacePrefix($prefix, $ns){
+  function addPrefix($prefix, $ns){
     $this->prefixes[$prefix] = $ns;
     $this->LDPath->setPrefixes($this->prefixes);
   }
@@ -53,15 +58,47 @@ class RafflesStore {
     $this->LDPath->setPrefixes($prefixes);
   }
 
+
   function load($descriptions){
     $count = array(
       's' => 0,
       'p' => 0,
       'o' => 0
     );
-    $lineNumbers = $this->DescriptionStore->insertDescriptions($descriptions);
+    $alreadyStoredDescriptions=array();
+    $newDescriptions = array();
     foreach($descriptions as $s => $props){
-      $lineNumber = $lineNumbers[$s];     
+      if($id = $this->Index->getSubject($s)){
+        $alreadyStoredDescriptions[$s]=$id;
+      } else {
+        $newDescriptions[$s] = $props;
+      }
+    }
+    
+    $existingIDs = array_values($alreadyStoredDescriptions);
+    $existingDescriptions = $this->describeIDs($existingIDs);
+
+    $mergedDescriptions = merge($descriptions,$existingDescriptions);
+    $replacements = array();
+    foreach($existingDescriptions as $e_uri => $e_props){
+      $id = $alreadyStoredDescriptions[$e_uri];
+      $replacements[$id] = array($e_uri => $mergedDescriptions[$e_uri]);
+    }
+    $lineNumbers = $this->DescriptionStore->insertDescriptions($newDescriptions);
+    $lineNumbers = array_merge($lineNumbers, $alreadyStoredDescriptions);
+    $this->DescriptionStore->replaceDescriptions($replacements);
+    gc_enable();
+    gc_collect_cycles();
+    gc_disable();
+    foreach($descriptions as $s => $props){
+      if(isset($lineNumbers[$s])){
+        $lineNumber = $lineNumbers[$s];
+      } else if(isset($alreadyStoredDescriptions[$s])) {
+        $lineNumber = $alreadyStoredDescriptions[$s]; 
+      } else {
+        throw new Exception($s.' somehow not indexed');
+      }
+      $lineNumber = $lineNumbers[$s];
       $this->Index->addSubject($s, $lineNumber);
       $count['s']++;
 
@@ -81,8 +118,11 @@ class RafflesStore {
       foreach($props as $p => $objs){
         $count['p']++;
         foreach($objs as $obj){
-          if(!$this->indexPredicates OR in_array($p, $this->indexPredicates)) $this->Index->addPredicateObject($p, $obj['value'], $lineNumber);
+          if(!$this->indexPredicates OR in_array($p, $this->indexPredicates)){
+            $this->Index->addPredicateObject($p, $obj['value'], $lineNumber);
+          }
           $count['o']++;
+          
         }
       }
     }
@@ -102,16 +142,66 @@ class RafflesStore {
   }
 
   function loadData($data){
+    require_once 'arc/ARC2.php';
+    $parser = ARC2::getRDFParser();
+    $parser->parse('',$data);
+    return $this->load($parser->getSimpleIndex(0));
     $graph = new EasyRdf_Graph();
     $graph->parse($data);
     return $this->load($graph->toArray());
   }
 
+  function loadDataFile($filename){
+    require_once 'streamingparsers.php';
+    $extension = array_pop(explode('.',$filename));
+    switch($extension){
+      case 'ttl':
+      case 'nt':
+      case 'n3':
+        $parser = new StreamingTurtleParser();
+        break;
+      case 'rdf':
+      case 'xml':
+        $parser = new StreamingRDFXMLParser();
+        break;
+      default:
+        throw Exception("No Handler for Data File extension: $extension");
+        break;
+    }
+    $callback = array(&$this, 'load');
+    $parser->setCallback($callback);
+    $parser->parse($filename);
+    $leftovers = $parser->getSimpleIndex(0);
+    $this->load($leftovers);
+  }
+
+  function createHierarchicalIndex(){
+    $ids = $this->Index->getAll();
+    $batches = array_chunk($ids, 50);
+    foreach($batches as $batch){
+      $descriptions = $this->describeIDs($batch);
+      foreach($descriptions as $uri => $props){
+        foreach($props as $p => $objs){
+          foreach($objs as $o){
+            if(isset($o['type']) AND ($o['type']=='uri' OR $o['type']=='bnode')){
+              $this->HierarchicalIndex->addTriple($uri,$p,$o['value']);
+            }
+          }
+        }
+      }
+    }
+  }
+
+  function relatedUris($uri){
+    return $this->HierarchicalIndex->getRelatedByURI($uri);
+  }
+  
   function getFacets($p){
     $propertyIndex = $this->Index->getPredicateValues($p);
     foreach($propertyIndex as $o => $lns){
       $propertyIndex[$o] = count($lns);
     }
+    arsort($propertyIndex);
     return $propertyIndex;
   }
 
@@ -121,13 +211,34 @@ class RafflesStore {
 
   function search($o_text, $property=false){
     $ids = $this->Index->searchObject($o_text,$property);
+    $this->_lastSet = $ids;
     return $this->DescriptionStore->getDescriptionsByIDs($ids);
   }
 
   function query($path, $limit=50,$offset=0){
     $triples = $this->LDPath->parse($path);
     $ids = $this->Index->query($triples);
+    $this->_lastSet = $ids;
     return $this->DescriptionStore->getDescriptionsByIDs(array_slice((array)$ids, $offset, $limit));
+  }
+
+  function getFacetsForLastQuery(){
+    $po =  $this->Index->filterPredicateObjectIndex($this->_lastSet);
+    $p_index = array();
+    foreach($po as $p => $os){
+      $p_index[$p] = count(array_keys($os));
+      foreach($os as $o => $ids){
+        $o = strval($o);
+        $po[$p][$o] = count($ids);
+      }
+      arsort($po[$p]);
+      $po[$p] = array_slice($po[$p],0,15,true);
+    }
+    asort($p_index);
+    foreach($p_index as $p => $count){
+      $p_index[$p] = $po[$p];
+    }
+    return $p_index;
   }
 
   function distance($uri, $km=30){
@@ -137,6 +248,7 @@ class RafflesStore {
       foreach($ids_by_distance as $distance => $d_ids){
         array_splice($ids,0,0,$d_ids);
       }
+      $this->_lastSet = $ids;
       return $this->describeIDs($ids);
     } else {
       return array();
@@ -144,14 +256,18 @@ class RafflesStore {
   }
 
   function __destruct() {
-    foreach($this->Index->po as $p => $o_ids){
-      if(is_array($o_ids)){
-        $filename = $this->dirname . DIRECTORY_SEPARATOR . 'index_po_' .urlencode($p);
-        file_put_contents($filename, serialize($o_ids), LOCK_EX);
-        $this->Index->po[$p] = $filename;
+    $index_file_name = $this->dirname . DIRECTORY_SEPARATOR .'index';
+    $descriptions_file_name = $this->dirname . DIRECTORY_SEPARATOR .'descriptions';
+    if(!is_file($index_file_name) OR is_file($descriptions_file_name) AND filemtime($descriptions_file_name) > filemtime($index_file_name) ){
+      foreach($this->Index->po as $p => $o_ids){
+        if(is_array($o_ids)){
+          $filename = $this->dirname . DIRECTORY_SEPARATOR . 'index_po_' .urlencode($p);
+          file_put_contents($filename, serialize($o_ids), LOCK_EX);
+          $this->Index->po[$p] = $filename;
+        }
       }
+      file_put_contents($index_file_name, serialize($this->Index), LOCK_EX);
     }
-    file_put_contents($this->dirname . DIRECTORY_SEPARATOR .'index', serialize($this->Index), LOCK_EX);
   }
 
   function reset(){
